Move plugin_manager fixture to root conftest
