+++ tests/conftest.py
@@ -7,6 +7,8 @@ from pytest_mock import MockerFixture
 
 import conda
 from conda.base.context import context, reset_context
+from conda.plugins.hookspec import CondaSpecs
+from conda.plugins.manager import CondaPluginManager
 
 from . import http_test_server
 
@@ -83,3 +85,11 @@ def do_not_register_envs(monkeypatch):
 def do_not_notify_outdated_conda(monkeypatch):
     """Do not notify about outdated conda during tests"""
     monkeypatch.setenv("CONDA_NOTIFY_OUTDATED_CONDA", "false")
+
+
+@pytest.fixture
+def plugin_manager(mocker) -> CondaPluginManager:
+    pm = CondaPluginManager()
+    pm.add_hookspecs(CondaSpecs)
+    mocker.patch("conda.plugins.manager.get_plugin_manager", return_value=pm)
+    return pm
